+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog of threedigrid-builder
 
 - Added connection node start and end id to lines.
 
+- Handle non-ASCII characters in gridadmin HDF5 output.
+
 
 0.15.0 (2022-01-05)
 -------------------